important refactoring code (part 1)

+++ pipeline/clustering/enrichment.py
@@ -6,7 +6,7 @@ warnings.filterwarnings("ignore")
 import os
 import argparse
 from pathlib import Path
-from utils.stdout import Section
+from utils.stdout import print_task, print_info, print_warning
 
 import re
 
@@ -83,14 +83,6 @@ parser.add_argument(
     help="file containing NCBI gene-to-GO terms (cannot be used with argument --annotations)"
 )
 
-parser.add_argument(
-    "-v", "--verbose",
-    dest="verbose",
-    required=False,
-    action="store_true",
-    help="display additional information"
-)
-
 args = parser.parse_args()
 
 if not Path(os.path.dirname(args.outfile)).exists():
@@ -103,12 +95,9 @@ elif args.go is not None and args.annotations is not None:
 else:
     annotations_alias_type = "geneid" if args.go else "MGI"
 
-section = Section(verbose = args.verbose)
 genesynonyms = GeneSynonyms()
 
-print(f"Loading data...")
-
-section("Loading background gene set...", reset=True)
+print_task("background gene set loading")
 
 population_ids = read_geneset(args.population)
 population_ids = genesynonyms.sequence_standardization(
@@ -123,7 +112,7 @@ if annotations_alias_type == "geneid":
 if None in population_ids:
     population_ids.remove(None)
 
-section("Loading study gene sets...")
+print_task("study gene sets loading")
 
 study_ids = dict()
 for _study_file in args.study:
@@ -141,7 +130,7 @@ for _study_file in args.study:
         _study_ids.remove(None)
     study_ids[os.path.basename(_study_file).rsplit(".", maxsplit=1)[0]] = _study_ids
 
-section("Loading gene ontologies...")
+print_task("gene ontologies loading")
 
 go_dag = GODag(args.go)
 
@@ -166,7 +155,7 @@ with open(args.go, "r") as go_reader:
             else:
                 continue
 
-section("Loading gene-to-go associations...")
+print_task("gene-to-go associations loading")
 
 if args.gene2go:
     annotations = Gene2GoReader(args.gene2go, taxids=[10090])
@@ -174,11 +163,10 @@ else:
     annotations = GafReader(args.annotations)
 associations = annotations.get_ns2assc()
 
-if args.verbose:
-    for namespace, geneid2go in associations.items():
-        print(f"{namespace} {len(geneid2go):,} annotated mouse genes")
+for namespace, geneid2go in associations.items():
+    print_info(f"{namespace} {len(geneid2go):,} annotated mouse genes")
 
-print(f"Gene Ontology Enrichment Analysis...")
+print_task("gene ontology enrichment analysis")
 
 goea = GOEnrichmentStudyNS(
     pop=population_ids,
@@ -194,20 +182,23 @@ for cluster, genes in study_ids.items():
     _goea_significant_results = [result for result in _goea_all_results if result.p_fdr_bh < 0.05]
     goea.wr_xlsx(f"{os.path.dirname(args.outfile)}/{cluster}", _goea_significant_results)
 
-print(f"Saving results...")
+print_task("data saving")
 
 with ExcelWriter(args.outfile) as xlsx_writer:
     for cluster in study_ids.keys():
         xlsx_infile = f"{os.path.dirname(args.outfile)}/{cluster}"
-        goea_results = read_excel(xlsx_infile, sheet_name=0)
-        for index, row in goea_results.iterrows():
-            _go = row["GO"]
-            if _go in go_definitions:
-                goea_results.at[index, "definition"] = go_definitions[_go]
-        column_names = list(goea_results.columns)
-        idx_study_items, idx_definition = column_names.index("study_items"), column_names.index("definition")
-        column_names[idx_definition], column_names[idx_study_items] = column_names[idx_study_items], column_names[idx_definition]
-        goea_results = goea_results[column_names]
-        goea_results.sort_values(by="p_fdr_bh", axis=0, ascending=True)
-        goea_results.to_excel(xlsx_writer, sheet_name=cluster)
-        os.remove(xlsx_infile)
+        if os.path.isfile(xlsx_infile):
+            goea_results = read_excel(xlsx_infile, sheet_name=0)
+            for index, row in goea_results.iterrows():
+                _go = row["GO"]
+                if _go in go_definitions:
+                    goea_results.at[index, "definition"] = go_definitions[_go]
+            column_names = list(goea_results.columns)
+            idx_study_items, idx_definition = column_names.index("study_items"), column_names.index("definition")
+            column_names[idx_definition], column_names[idx_study_items] = column_names[idx_study_items], column_names[idx_definition]
+            goea_results = goea_results[column_names]
+            goea_results.sort_values(by="p_fdr_bh", axis=0, ascending=True)
+            goea_results.to_excel(xlsx_writer, sheet_name=cluster)
+            os.remove(xlsx_infile)
+        else:
+            print_warning(f"{xlsx_infile} not found")
